+++ python/discover/background.py
@@ -22,14 +22,8 @@ def estimate_background(events, strata=None):
 
         mu = maxent.fit(rowValues, rowWeights, colValues, colWeights)
         nRows = len(rowValues)
-        eA = numpy.dot(numpy.exp((mu[:nRows] / rowWeights)[:, numpy.newaxis]),
-                       numpy.exp((mu[nRows:] / colWeights)[numpy.newaxis]))
+        eA = numpy.outer(numpy.exp(mu[:nRows]), numpy.exp(mu[nRows:]))
         P = 1.0 / (eA + 1)
-
-        if P.max() > 1:
-            import warnings
-            warnings.warn("Some background estimates are greater than 1")
-            #P[P > 1] = 1
 
         return P[rowValues.searchsorted(rowSums)[:, numpy.newaxis],
                  colValues.searchsorted(colSums)[numpy.newaxis]]